+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add translations to content purpose supergroup titles
+
 # 0.8.0
 
 * Adds `GovukDocumentTypes.supergroup_subgroups` as a way to retrieve